Move Schema classes to Types folder
